初步拆解 原allin1f中框架部分以及共用实例部分。 依赖aiagent完成自主分析历史内容进行改名的动作。
